+++ Editor/Change version/Class's/BuildPhaseTemplate.cs
@@ -1,9 +1,0 @@
-namespace Cobilas.Unity.Editor.Utility.ChangeVersion {
-    public sealed class BuildPhaseTemplate : BaseBuildPhaseTemplate {
-        public override string[] GetTemplates()
-            => new string[] {
-                "alpha",
-                "beta"
-            };
-    }
-}